Reboot handling in wizard

+++ patchbox/modules/boot/cli.py
@@ -1,7 +1,7 @@
 import subprocess
 import click
 import os
-from patchbox.utils import do_group_menu, do_ensure_param, do_go_back_if_ineractive, do_yesno
+from patchbox.utils import do_group_menu, do_ensure_param, do_go_back_if_ineractive, do_yesno, do_msgbox
 
 
 @click.group(invoke_without_command=True)
@@ -30,11 +30,12 @@ def environment(ctx, option):
     elif option == 'console autologin':
         subprocess.call([os.path.join(os.path.dirname(os.path.realpath(__file__)), 'scripts/set_boot_to_console.sh'), 'autologin'])
 
-    if ctx.meta.get('interactive'):
-        no, output = do_yesno('A system restart is required to activate the new behavior. Reboot now?')
+    ctx.meta['reboot_required'] = True
+    if ctx.meta.get('interactive') and not ctx.meta.get('wizard'):
+        no, output = do_yesno('A system restart is required to activate the new boot behavior. Reboot now?')
         if not no:
             subprocess.call('reboot')
     else:
-        click.echo('A system restart is required to activate the new behavior. Do so by executing `sudo reboot`.')
+        click.echo('A system restart is required to activate the new boot behavior. You may do so by executing `sudo reboot`.')
 
     do_go_back_if_ineractive(silent=True)
